requirements.txt
dj-database-url==0.5.0
django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
docutils==0.17.1
geocoder==1.38.1
gunicorn==20.1.0
ip2geotools==0.1.5
IP2Location==8.5.1
jsonschema==3.2.0
packaging==20.9
pip-review==1.1.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyquery==1.4.3
ratelim==0.1.6
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.3
webencodings==0.5.1
whitenoise==5.2.0
